loadAllMinifiers
