+++ lib/createWebsite/makeRealWorldTable.js
@@ -3,8 +3,9 @@ import { join } from 'node:path';
 
 import prettyMilliseconds from 'pretty-ms';
 
+import { minifiers } from '../loaders/loadAllMinifiers.js';
 import { loadAllRealWorldTests } from '../loaders/loadAllRealWorldTests.js';
-import { minifiers, minify } from '../minify.js';
+import { minify } from '../minify.js';
 
 const __dirname = import.meta.dirname;
 const ERROR = 'ERROR';
