+++ lib/minify.js
@@ -1,26 +1,7 @@
 import fs from "node:fs";
 import path from "node:path";
-import cleanCss from "./minifiers/clean-css.js";
-import csslop from "./minifiers/csslop.js";
-import csskit from "./minifiers/csskit.js";
-import cssnano from "./minifiers/cssnano.js";
-import csso from "./minifiers/csso.js";
-import esbuild from "./minifiers/esbuild.js";
-import lightningcss from "./minifiers/lightningcss.js";
-import sass from "./minifiers/sass.js";
 
-const registry = new Map([
-  ["clean-css", cleanCss],
-  ["csskit", csskit],
-  ["csslop", csslop],
-  ["cssnano", cssnano],
-  ["csso", csso],
-  ["esbuild", esbuild],
-  ["lightningcss", lightningcss],
-  ["sass", sass],
-]);
-
-export const minifiers = [...registry.keys()];
+import { minifiers, registry } from "./loaders/loadAllMinifiers.js";
 
 export async function minify(name, source) {
   const fn = registry.get(name);
